requirements.txt
streamlit
openai
pypdf2
python-dotenv
